Moving Platforms
Kind of rough but it's a start. You need to make a Moving Platform
object that is a trigger and one that is not, and make them both have
the same starting position, velocities, etc. to make the whole thing
work. There are a few bugs- if the player gets hit or the platform
touches an enemy (or sometimes an enemy bullet, but this is infrequent),
the player will stop moving with the platform unless they jump off and
back on to it.